Move DebouncedColorPicker to `@/components/layout/BaseColors`
