+++ src/components/layout/BaseColors/index.ts
@@ -1,1 +1,2 @@
 export * from './ColorListItem';
+export * from './DebouncedColorPicker';
